requirements.txt
streamlit
pandas
scikit-learn
selenium
fpdf
python-docx
streamlit-javascript
